Addressing Issues from first PR
A change in the configuration of the sinv and cmpbqm input yamls. A 'platform' field was added to these yamls as a way the tool can differentiate between platforms.

Corrected some typos and changed some comments to be more accurate.

+++ src/obs_inv_utils/nceplibs_bufr_cmd_handler.py
@@ -6,7 +6,7 @@ import os
 from pathlib import Path
 import shutil
 import uuid
-
+import yaml
 import pandas as pd
 from pandas import DataFrame
 import pathlib
@@ -23,7 +23,6 @@ from obs_inv_utils import subprocess_cmd_handler as sch
 from obs_inv_utils.subprocess_cmd_handler import SubprocessCmd
 from obs_inv_utils import nceplibs_cmds as nc_cmds
 
-import yaml
 
 CALLING_DIR = pathlib.Path(__file__).parent.resolve()
 TMP_OBS_DATA_DIR = 'tmp_obs_data'
@@ -106,6 +105,7 @@ class ObsBufrFileMetaHandler(object):
     config_data: dict = field(default_factory=str, init=False)
     s3_bucket: str = field(default_factory=str, init=False)
     s3_prefix: str = field(default_factory=str, init=False)
+    platform:  str = field(default_factory=str, init=False)
 
 
     def __post_init__(self):
@@ -128,14 +128,16 @@ class ObsBufrFileMetaHandler(object):
             self.date_range.start,
             self.date_range.end
         )
-        
-        bucket = self.meta_config.s3_bucket
-        prefix = self.meta_config.s3_prefix
+
+        # Delete?
+        #bucket = self.meta_config.s3_bucket
+        #prefix = self.meta_config.s3_prefix
+        #platform = self.meta_config.platform
 
 
-        # Because s3 bucket bufr files require download and new work_dir and discover files do not they need to be handled differently. 
-        if bucket == 'noaa-reanalyses-pds': 
-            print(f'Running get_bufr_file_meta for NOAA S3 Clean Bucket (aws_s3) {self.s3_bucket}')
+        # Added 'platform' to config_handler - now a required field in the input yaml files for get-obs-count-meta-sinv and get-obs-count-meta-cmpbqm  
+        if self.meta_config.platform == 'aws_s3': 
+            print(f'Running get_bufr_file_meta for AWS S3 Bucket: {self.meta_config.platform}')
 
             temp_uuid = str(uuid.uuid4())
 
@@ -160,12 +162,10 @@ class ObsBufrFileMetaHandler(object):
                     os.remove( saved_filename )
                     shutil.rmtree( work_dir )
               
-        elif self.s3_bucket == None or self.s3_bucket == '':
-            print(f'Running get_bufr_file_meta for NASA Discover {self.s3_bucket}')
+        elif self.meta_config.platform == 'discover':
+            print(f'Running get_bufr_file_meta for NASA Discover {self.meta_config.platform}')
 
-            temp_uuid = str(uuid.uuid4())
-
-            work_dir = os.path.join(self.meta_config.work_dir, temp_uuid)
+            work_dir = os.path.join(self.meta_config.work_dir)
 
             for idx, bufr_file in inventory_bufr_files.iterrows():
                 file_downloaded = False
@@ -213,10 +213,17 @@ class ObsPrepBufrFileMetaHandler(object):
     meta_config: ObsMetaCMPBQMConfig
     prepbufr_files: list = field(default_factory=list, init=False)
     date_range: DateRange = field(init=False)
+    # additional fields required in order to differentiate between aws and discover
+    config_data: dict = field(default_factory=str, init=False)
+    s3_bucket: str = field(default_factory=str, init=False)
+    s3_prefix: str = field(default_factory=str, init=False)
+    platform:  str = field(default_factory=str, init=False)
+
 
     def __post_init__(self):
         self.date_range = self.meta_config.get_date_range()
         self.prepbufr_files = self.meta_config.get_prepbufr_file_list()
+        self.config_data = self.meta_config.load()
 
     def __repr__(self):
         """
@@ -234,27 +241,63 @@ class ObsPrepBufrFileMetaHandler(object):
             self.date_range.end
         )
 
-        temp_uuid = str(uuid.uuid4())
+        # Delete?
+        #bucket = self.meta_config.s3_bucket
+        #prefix = self.meta_config.s3_prefix
+        #platform = self.meta_config.platform
 
-        work_dir = os.path.join(self.meta_config.work_dir, temp_uuid)
+        # Added 'platform' to config_handler - now a required field in the input yaml files for get-obs-count-meta-sinv and get-obs-count-meta-cmpbqm  
+        if self.meta_config.platform == 'aws_s3':
+            print(f'Running get_bufr_file_meta for S3 Bucket: {self.meta_config.platform}')
 
-        print(f'inventory_prepbufr_files: {inventory_prepbufr_files}')
-        print(f'scrub_files: {self.meta_config.scrub_files}')
-        for idx, prepbufr_file in inventory_prepbufr_files.iterrows():
-            print(
-               f'bufr_file: {prepbufr_file}')
+            temp_uuid = str(uuid.uuid4())
 
-            saved_filename = download_bufr_file_from_s3(work_dir, prepbufr_file)
+            work_dir = os.path.join(self.meta_config.work_dir, temp_uuid)
 
-            if saved_filename is None:
-                continue
+            print(f'inventory_prepbufr_files: {inventory_prepbufr_files}')
+            print(f'scrub_files: {self.meta_config.scrub_files}')
+            for idx, prepbufr_file in inventory_prepbufr_files.iterrows():
+                file_downloaded = False
+                print(
+                   f'bufr_file: {prepbufr_file}')
+                
+                saved_filename = download_bufr_file_from_s3(work_dir, prepbufr_file)
+  
+                if saved_filename is None:
+                    continue
 
-            self.get_obs_counts_with_cmpbqm(saved_filename, prepbufr_file)
+                self.get_obs_counts_with_cmpbqm(saved_filename, prepbufr_file)
 
-            # clean up files
-            if self.meta_config.scrub_files:
-                shutil.rmtree( work_dir )
+                # clean up files
+                if self.meta_config.scrub_files:
+                    shutil.rmtree( work_dir )
+                    
+        elif self.meta_config.platform == 'discover':
+            print(f'Running get_bufr_file_meta for NASA Discover {self.meta_config.platform}')
 
+            work_dir = os.path.join(self.meta_config.work_dir)
+
+            print(f'inventory_prepbufr_files: {inventory_prepbufr_files}')
+            print(f'scrub_files: {self.meta_config.scrub_files}')
+            for idx, prepbufr_file in inventory_prepbufr_files.iterrows():
+                file_downloaded = False
+                print(
+                   f'bufr_file: {prepbufr_file}')
+ 
+                try:
+                    saved_filename = prepbufr_file['full_path'] 
+                    print(f'saved_filename: {saved_filename}')
+                except Exception as err:
+                    msg = f'\'saved_filename\' line 147 - err: {err}'
+                    raise ValueError(msg) from err 
+
+                if saved_filename is None:
+                    continue
+
+                self.get_obs_counts_with_cmpbqm(saved_filename, prepbufr_file)
+
+
+    
     def get_obs_counts_with_cmpbqm(self, filename, prepbufr_file):
         args = [filename]
         cmd = sch.SubprocessCmdHandler(
